refactor : 리뷰 반영 1차 리팩토링

+++ src/main/java/lotto/domain/PrizeSummary.java
@@ -10,9 +10,9 @@ public class PrizeSummary {
         this.prizeSummary = prizeSummary;
     }
 
-    public int calculatePrizeAmount() {
+    public long calculatePrizeAmount() {
         return prizeSummary.entrySet().stream()
-                .mapToInt(prizeDetail -> prizeDetail.getKey().calculateRankByCount(prizeDetail.getValue()))
+                .mapToLong(prizeDetail -> prizeDetail.getKey().calculateRankByCount(prizeDetail.getValue()))
                 .sum();
     }
